+++ include/rcdl/rcdl.h
@@ -35,6 +35,7 @@
 #include "rcdl/tasks/ocr.h"
 #include "rcdl/tasks/face.h"
 #include "rcdl/tasks/features.h"
+#include "rcdl/tasks/superres.h"
 #include "rcdl/tracks/byte_tracker.h"
 #include "rcdl/tracks/reid.h"
 #include "rcdl/backend/engine_pool.h"